backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
adafruit-circuitpython-neopixel
rpi-lgpio
Adafruit-Blinka-Raspberry-Pi5-Neopixel